Add settickets and getpinfo
This commit adds the settickets and getpinfo functions to xv6. The settickets function can be
used to set the number of tickets of the calling process. This will be used by the stride
scheduler to determine how often the process should be scheduled. The getpinfo function will
build a pstats structure that contains the inuse value, pid, ticket number, and number of ticks
for every process in xv6.