Add ticket infrastructure to process struct
This commit adds all of the infrastructure required for the stride
scheduler to be implimented. Processes will now be created with
10 tickets and pass their ticket value down to their children. The
settickets() function has also been modified to properly set the
tickets value of the calling process. Considerations have also been
made for processes waking up and new processes made so that they will
not start with a pass value. This will prevent them from starving
all of the existing processes.