Impliment Stride Scheduler
In this commit the stride scheduler was implimented. This
was done by cycling through all RUNNABLE processes and
selecting the process with the lowest pass value. This
value is saved for later use and the selected process is
switched into. The selected process also has its pass value
increased by its stride value.